cleanup and refactoring of tests
